Moved all CSS files into a CSS folder
