+++ tiles01/src/CSS/Card.css
@@ -2,8 +2,6 @@
     width: 10rem;
     height: auto;
     margin: 5% 2%;
-    /* width:15%;
-     */
 }
 
 .card:hover{
@@ -16,7 +14,4 @@
     box-shadow: 0px 12px 17px 2px rgba(0, 0, 0, 0.14),
             0px 5px 22px 4px rgba(0, 0, 0, 0.12),
             0px 7px 8px -4px rgba(0, 0, 0, 0.2);
-    /*box-shadow: 5px 12px rgba(0, 0, 0, 0.2);*/
-    /*border: 0.375rem rgba(108, 143, 143, 0.589) solid;*/
-    /*width: 11rem;*/
 }
