+++ tiles01/src/Components/Logo.jsx
@@ -1,6 +1,4 @@
 import React from 'react';
-//import { Link } from "react-router-dom";
-
 
 function Logo(props) {
     const styling ={
