+++ tiles01/src/Components/TopContainer.jsx
@@ -1,7 +1,5 @@
 import React from 'react';
-//import "../App.css";
-import "../TopContainer.css";
-//import { Link } from "react-router-dom";
+import "../CSS/TopContainer.css";
 import Logo from "./Logo";
