+++ tiles01/src/Pages/AnimationPage.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import { Link } from 'react-router-dom';
 import TopContainer from '../Components/TopContainer'
 import LinkButton from '../Components/LinkButton'
